Add files via upload
These files, minus the final.zip, are the main files needed to run the labs. The CSVparser was given to us so we could use the excel file and maintain focus on the algorithm portion of the class. The Final.zip is the folder that contains a reflection of the course's work.